daemon: move errors-related tests from response_test.go to errors_test.go
