+++ daemon/response_test.go
@@ -21,7 +21,6 @@ package daemon_test
 
 import (
 	"encoding/json"
-	"errors"
 	"fmt"
 	"io/ioutil"
 	"net/http"
@@ -31,11 +30,7 @@ import (
 
 	"gopkg.in/check.v1"
 
-	"github.com/snapcore/snapd/client"
 	"github.com/snapcore/snapd/daemon"
-	"github.com/snapcore/snapd/overlord/snapstate"
-	"github.com/snapcore/snapd/snap"
-	"github.com/snapcore/snapd/store"
 )
 
 type responseSuite struct{}
@@ -113,129 +108,4 @@ func (s *responseSuite) TestRespJSONWithNullResult(c *check.C) {
 	data, err := json.Marshal(rj)
 	c.Assert(err, check.IsNil)
 	c.Check(string(data), check.Equals, `{"type":"","status-code":0,"status":"","result":null}`)
-}
-
-func (responseSuite) TestErrorResponderPrintfsWithArgs(c *check.C) {
-	teapot := daemon.MakeErrorResponder(418)
-
-	rec := httptest.NewRecorder()
-	rsp := teapot("system memory below %d%%.", 1)
-	req, err := http.NewRequest("GET", "", nil)
-	c.Assert(err, check.IsNil)
-	rsp.ServeHTTP(rec, req)
-
-	var v struct{ Result daemon.ErrorResult }
-	c.Assert(json.NewDecoder(rec.Body).Decode(&v), check.IsNil)
-
-	c.Check(v.Result.Message, check.Equals, "system memory below 1%.")
-}
-
-func (responseSuite) TestErrorResponderDoesNotPrintfAlways(c *check.C) {
-	teapot := daemon.MakeErrorResponder(418)
-
-	rec := httptest.NewRecorder()
-	rsp := teapot("system memory below 1%.")
-	req, err := http.NewRequest("GET", "", nil)
-	c.Assert(err, check.IsNil)
-	rsp.ServeHTTP(rec, req)
-
-	var v struct{ Result daemon.ErrorResult }
-	c.Assert(json.NewDecoder(rec.Body).Decode(&v), check.IsNil)
-
-	c.Check(v.Result.Message, check.Equals, "system memory below 1%.")
-}
-
-type fakeNetError struct {
-	message   string
-	timeout   bool
-	temporary bool
-}
-
-func (e fakeNetError) Error() string   { return e.message }
-func (e fakeNetError) Timeout() bool   { return e.timeout }
-func (e fakeNetError) Temporary() bool { return e.temporary }
-
-func (s *responseSuite) TestErrToResponse(c *check.C) {
-	aie := &snap.AlreadyInstalledError{Snap: "foo"}
-	nie := &snap.NotInstalledError{Snap: "foo"}
-	cce := &snapstate.ChangeConflictError{Snap: "foo"}
-	ndme := &snapstate.SnapNeedsDevModeError{Snap: "foo"}
-	nc := &snapstate.SnapNotClassicError{Snap: "foo"}
-	nce := &snapstate.SnapNeedsClassicError{Snap: "foo"}
-	ncse := &snapstate.SnapNeedsClassicSystemError{Snap: "foo"}
-	netoe := fakeNetError{message: "other"}
-	nettoute := fakeNetError{message: "timeout", timeout: true}
-	nettmpe := fakeNetError{message: "temp", temporary: true}
-
-	e := errors.New("other error")
-
-	sa1e := &store.SnapActionError{Refresh: map[string]error{"foo": store.ErrSnapNotFound}}
-	sa2e := &store.SnapActionError{Refresh: map[string]error{
-		"foo": store.ErrSnapNotFound,
-		"bar": store.ErrSnapNotFound,
-	}}
-	saOe := &store.SnapActionError{Other: []error{e}}
-	// this one can't happen (but fun to test):
-	saXe := &store.SnapActionError{Refresh: map[string]error{"foo": sa1e}}
-
-	makeErrorRsp := func(kind client.ErrorKind, err error, value interface{}) *daemon.APIError {
-		return &daemon.APIError{
-			Status:  400,
-			Message: err.Error(),
-			Kind:    kind,
-			Value:   value,
-		}
-	}
-
-	tests := []struct {
-		err         error
-		expectedRsp daemon.Response
-	}{
-		{store.ErrSnapNotFound, daemon.SnapNotFound("foo", store.ErrSnapNotFound)},
-		{store.ErrNoUpdateAvailable, makeErrorRsp(client.ErrorKindSnapNoUpdateAvailable, store.ErrNoUpdateAvailable, "")},
-		{store.ErrLocalSnap, makeErrorRsp(client.ErrorKindSnapLocal, store.ErrLocalSnap, "")},
-		{aie, makeErrorRsp(client.ErrorKindSnapAlreadyInstalled, aie, "foo")},
-		{nie, makeErrorRsp(client.ErrorKindSnapNotInstalled, nie, "foo")},
-		{ndme, makeErrorRsp(client.ErrorKindSnapNeedsDevMode, ndme, "foo")},
-		{nc, makeErrorRsp(client.ErrorKindSnapNotClassic, nc, "foo")},
-		{nce, makeErrorRsp(client.ErrorKindSnapNeedsClassic, nce, "foo")},
-		{ncse, makeErrorRsp(client.ErrorKindSnapNeedsClassicSystem, ncse, "foo")},
-		{cce, daemon.SnapChangeConflict(cce)},
-		{nettoute, makeErrorRsp(client.ErrorKindNetworkTimeout, nettoute, "")},
-		{netoe, daemon.BadRequest("ERR: %v", netoe)},
-		{nettmpe, daemon.BadRequest("ERR: %v", nettmpe)},
-		{e, daemon.BadRequest("ERR: %v", e)},
-
-		// action error unwrapping:
-		{sa1e, daemon.SnapNotFound("foo", store.ErrSnapNotFound)},
-		{saXe, daemon.SnapNotFound("foo", store.ErrSnapNotFound)},
-		// action errors, unwrapped:
-		{sa2e, daemon.BadRequest(`ERR: cannot refresh: snap not found: "bar", "foo"`)},
-		{saOe, daemon.BadRequest("ERR: cannot refresh, install, or download: other error")},
-	}
-
-	for _, t := range tests {
-		com := check.Commentf("%v", t.err)
-		rspe := daemon.ErrToResponse(t.err, []string{"foo"}, daemon.BadRequest, "%s: %v", "ERR")
-		c.Check(rspe, check.DeepEquals, t.expectedRsp, com)
-	}
-}
-
-func (s *responseSuite) TestErrToResponseInsufficentSpace(c *check.C) {
-	err := &snapstate.InsufficientSpaceError{
-		Snaps:      []string{"foo", "bar"},
-		ChangeKind: "some-change",
-		Path:       "/path",
-		Message:    "specific error msg",
-	}
-	rspe := daemon.ErrToResponse(err, nil, daemon.BadRequest, "%s: %v", "ERR")
-	c.Check(rspe, check.DeepEquals, &daemon.APIError{
-		Status:  507,
-		Message: "specific error msg",
-		Kind:    client.ErrorKindInsufficientDiskSpace,
-		Value: map[string]interface{}{
-			"snap-names":  []string{"foo", "bar"},
-			"change-kind": "some-change",
-		},
-	})
 }
